Update Prime_Time v2.0.py

+++ Prime_Time v2.0.py
@@ -44,7 +44,7 @@ def print_primes(l, u):
     if l % 2 == 0:
         for num in range(l+1, u+1, 2):
             prime = True
-            for i in range(3, int(math.sqrt(num)), 2):
+            for i in range(3, math.sqrt(num), 2):
                 if num % i == 0:
                     prime = False
                     break
@@ -53,7 +53,7 @@ def print_primes(l, u):
     else:
         for num in range(l, u+1, 2):
             prime = True
-            for i in range(3, int(math.sqrt(num)), 2):
+            for i in range(3, math.sqrt(num), 2):
                 if num % i == 0:
                     prime = False
                     break
